Move pipeline implementations to FeatureLib
