+++ Sources/Plasma/FeatureLib/CMakeLists.txt
@@ -15,8 +15,14 @@ add_subdirectory(pfConsoleCore)
 if(WIN32)
     add_subdirectory(pfCrashHandler)
 endif()
+if(PLASMA_PIPELINE STREQUAL "DirectX")
+    add_subdirectory(pfDXPipeline)
+endif()
 add_subdirectory(pfGameGUIMgr)
 add_subdirectory(pfGameScoreMgr)
+if(PLASMA_PIPELINE STREQUAL "OpenGL")
+    add_subdirectory(pfGLPipeline)
+endif()
 add_subdirectory(pfJournalBook)
 add_subdirectory(pfLocalizationMgr)
 add_subdirectory(pfMessage)
